pyproject.toml
[project]
name = "golftrolley-yolo"
version = "0.1.0"
description = "Self driving golf trolley"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
]
